api-chatx/app/main.py
from fastapi import FastAPI
import asyncpg
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI()

DATABASE_URL = os.getenv("DATABASE_URL")

@app.on_event("startup")
async def startup():
    app.state.pool = await asyncpg.create_pool(DATABASE_URL)

@app.on_event("shutdown")
async def shutdown():
    await app.state.pool.close()

@app.get("/webhook/menu")
async def get_menu():
    async with app.state.pool.acquire() as conn:
        result = await conn.fetchval("SELECT get_menu_json();")
        return result
